+++ src/storage/filesystem.py
@@ -7,6 +7,7 @@ from binascii import hexlify, unhexlify
 from os import listdir, makedirs, path, remove, rmdir
 from threading import RLock
 
+from debug import logger
 from paths import lookupAppdataFolder
 from storage import InventoryItem, InventoryStorage
 
@@ -162,7 +163,7 @@ class FilesystemInventory(InventoryStorage):
                     newInventory[streamNumber][hashId] = InventoryItem(
                         objectType, streamNumber, None, expiresTime, tag)
             except KeyError:
-                print "error loading %s" % (hexlify(hashId))
+                logger.warning('error loading %s', hexlify(hashId))
         self._inventory = newInventory
 #        for i, v in self._inventory.items():
 #            print "loaded stream: %s, %i items" % (i, len(v))
